Fundzin Backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
Automat==20.2.0
black==21.9b0
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
docutils==0.17.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonmerge==1.8.0
jsonschema==4.0.1
lockfile==0.12.2
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.21
pydoctor==21.9.1
PyJWT==1.7.1
pyrsistent==0.18.0
pytz==2021.1
regex==2021.9.30
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zope.interface==5.4.0
